pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "persona-studio"
version = "0.1.0"
description = "ETL pipeline that turns heterogeneous source material into a single markdown corpus for persona synthesis."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Human AI Agent Creator" }]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.6",
    "pypdf>=4.2",
    "python-docx>=1.1",
    "python-pptx>=0.6.23",
    "markdown-it-py>=3.0",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.9",
    "faster-whisper>=1.0",
    "yt-dlp>=2024.4.9",
    "youtube-transcript-api>=0.6.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
persona-studio = "persona_studio.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/persona_studio"]
omit = [
    "src/persona_studio/ingest/audio.py",
    "src/persona_studio/ingest/youtube.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.hatch.build.targets.wheel]
packages = ["src/persona_studio"]

[tool.hatch.build.targets.sdist]
include = ["src/persona_studio", "README.md", "LICENSE"]
